\name{asWrapTry-package}
\alias{asWrapTry-package}
\alias{asWrapTry}
\docType{package}
\title{
\packageTitle{asWrapTry}
}
\description{
\packageDescription{asWrapTry}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{asWrapTry}
\packageIndices{asWrapTry}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{asWrapTry}

Maintainer: \packageMaintainer{asWrapTry}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
